+++ packages/shared/src/enums/streams.enum.ts
@@ -20,6 +20,8 @@ export enum PostStream {
   POST_JOB_OUTBOX_FAILURE = 'post:job:outbox:failure',
   POST_CREATED = 'post-created',
   POST_DELETED = 'post-deleted',
+  COMMENT_CREATED = 'comment-created',
+  COMMENT_DELETED = 'comment-deleted',
 }
 
 export enum SocialStream {