+++ packages/outbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.9.45
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/shared@0.8.2
+  - @volontariapp/database@3.4.8
+  - @volontariapp/messaging@2.12.1
+
 ## 0.9.44
 
 ### Patch Changes